Create CCD config function

+++ firmware/Inc/experiment_ccd.h
@@ -7,7 +7,7 @@
 
 
 // Period config
-extern uint16_t ccd_config_poll_period_sec;  // [default: 15] configurable via command, extern; 0 to disable
+extern uint8_t ccd_config_poll_period_sec;  // [default: 15] configurable via command, extern; 0 to disable
 extern uint16_t ccd_config_stat_period_sec;  // [default: 120] configurable via command, extern; 0 to disable
 
 // Operational/functional config
